[Pro Fix] 오류 수정 (데미지 연산 오류 수정/영역SO 수치 오류 수정)
상세 - 데미지 연산 오류 수정
내부 데미지 데이터 타입 변경 (int -> float), 즉 내부 연산은 float으로 한 후, 최종 공격시, int로 올림처리

+++ GameProject/Assets/Game/0_Scripts/CardReletive/Effects/AttackEnemyEffect.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class AttackEnemyEffect : Effect
 {
-    [SerializeField] private int amount;
+    [SerializeField] private float amount;
     public override GameAction GetGameAction(EffectInfo effectInfo)
     {
         AttackEnemyGA attackEnemyGA = new(effectInfo.targetPoses, CalculateDamage(amount));
