+++ GameProject/Assets/Game/0_Scripts/CardReletive/GameActions/CardEffectRelative/ShieldBashGA.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class ShieldBashGA : GameAction
 {
-    public int Amount { get; private set; }
+    public float Amount { get; private set; }
     public List<Vector2Int> TargetPoses { get; private set; }
-    public ShieldBashGA(int amount, List<Vector2Int> targetPoses)
+    public ShieldBashGA(float amount, List<Vector2Int> targetPoses)
     {
         Amount = amount;
         TargetPoses = targetPoses;
