+++ GameProject/Assets/Game/0_Scripts/CardReletive/GameActions/CardEffectRelative/ShoulderBashGA.cs
@@ -7,10 +7,10 @@ public class ShoulderBashGA : GameAction
 {
     public int Distance { get; private set; }
     public int AttackDistance { get; private set; }
-    public int Damage { get; private set; }
+    public float Damage { get; private set; }
     public List<Vector2Int> TargetPoses { get; private set; }
 
-    public ShoulderBashGA(int distance, int attackDistance, int damage, List<Vector2Int> targetPoses)
+    public ShoulderBashGA(int distance, int attackDistance, float damage, List<Vector2Int> targetPoses)
     {
         this.Distance = distance;
         this.AttackDistance = attackDistance;
