+++ GameProject/Assets/Game/0_Scripts/CardReletive/GameActions/DealDamageGA.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class DealDamageGA : GameAction, IHaveCaster
 {
-    public int Amount { get; set; }
+    public float Amount { get; set; }
     public List<CombatantView> Targets { get; private set; }
 
     public CombatantView Caster { get; private set; }
     public DamageFormulaType FormulaType { get; set; }
 
-    public DealDamageGA(int amount, List<CombatantView> targets, CombatantView caster, DamageFormulaType formulaType = DamageFormulaType.Main)
+    public DealDamageGA(float amount, List<CombatantView> targets, CombatantView caster, DamageFormulaType formulaType = DamageFormulaType.Main)
     {
         Amount = amount;
         Targets = new(targets);
