+++ GameProject/Assets/Game/0_Scripts/CardReletive/GameActions/HealGA.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 
 public class HealGA : GameAction
 {
-    public int Amount { get; private set; }
+    public float Amount { get; private set; }
     public List<Vector2Int> TargetPoses { get; private set; }
     public List<CombatantView> Targets { get; private set; }
 
     //그리드 기반
-    public HealGA(int amount, List<Vector2Int> targetPoses)
+    public HealGA(float amount, List<Vector2Int> targetPoses)
     {
         Amount = amount;
         TargetPoses = targetPoses;
@@ -17,7 +17,7 @@ public class HealGA : GameAction
     }
 
     //대상 기반
-    public HealGA(int amount, List<CombatantView> target)
+    public HealGA(float amount, List<CombatantView> target)
     {
         Amount = amount;
         Targets = target;
